Trait move_vm_types::chain_state::ChainState

pub trait ChainState {
    fn deduct_gas(&mut self, amount: GasUnits<GasCarrier>) -> VMResult<()>;

    fn remaining_gas(&self) -> GasUnits<GasCarrier>;

    fn load_module(&self, module: &ModuleId) -> VMResult<Vec<u8>>;

    fn borrow_resource(
        &mut self, 
        ap: &AccessPath, 
        ty: &FatStructType
    ) -> VMResult<Option<&GlobalValue>>;

    fn move_resource_from(
        &mut self, 
        ap: &AccessPath, 
        ty: &FatStructType
    ) -> VMResult<Option<GlobalValue>>;

    fn publish_module(
        &mut self, 
        module_id: ModuleId, 
        module: Vec<u8>
    ) -> VMResult<()>;

    fn publish_resource(
        &mut self, 
        ap: &AccessPath, 
        g: (FatStructType, GlobalValue)
    ) -> VMResult<()>;

    fn exists_module(&self, key: &ModuleId) -> bool;

    fn emit_event(&mut self, event: ContractEvent);
}

Trait that describes what Move bytecode runtime expects from the Libra blockchain.

Required methods

fn deduct_gas(&mut self, amount: GasUnits<GasCarrier>) -> VMResult<()>

fn remaining_gas(&self) -> GasUnits<GasCarrier>

fn load_module(&self, module: &ModuleId) -> VMResult<Vec<u8>>

Get the serialized format of a CompiledModule from chain given a ModuleId.

fn borrow_resource(
&mut self,
ap: &AccessPath,
ty: &FatStructType
) -> VMResult<Option<&GlobalValue>>

Get a reference to a resource stored on chain.

fn move_resource_from(
&mut self,
ap: &AccessPath,
ty: &FatStructType
) -> VMResult<Option<GlobalValue>>

Transfer ownership of a resource stored on chain to the VM.

fn publish_module(
&mut self,
module_id: ModuleId,
module: Vec<u8>
) -> VMResult<()>

Publish a module to be stored on chain.

fn publish_resource(
&mut self,
ap: &AccessPath,
g: (FatStructType, GlobalValue)
) -> VMResult<()>

Publish a resource to be stored on chain.

fn exists_module(&self, key: &ModuleId) -> bool

Check if this module exists on chain.

fn emit_event(&mut self, event: ContractEvent)

Emit an event to the EventStore

Loading content...