Trait move_vm_types::gas::GasMeter

pub trait GasMeter {

    fn charge_simple_instr(
        &mut self, 
        instr: SimpleInstruction
    ) -> PartialVMResult<()>;
    fn charge_pop(&mut self, popped_val: impl ValueView) -> PartialVMResult<()>;
    fn charge_call(
        &mut self, 
        module_id: &ModuleId, 
        func_name: &str, 
        args: impl ExactSizeIterator<Item = impl ValueView>, 
        num_locals: NumArgs
    ) -> PartialVMResult<()>;
    fn charge_call_generic(
        &mut self, 
        module_id: &ModuleId, 
        func_name: &str, 
        ty_args: impl ExactSizeIterator<Item = impl TypeView>, 
        args: impl ExactSizeIterator<Item = impl ValueView>, 
        num_locals: NumArgs
    ) -> PartialVMResult<()>;
    fn charge_ld_const(&mut self, size: NumBytes) -> PartialVMResult<()>;
    fn charge_ld_const_after_deserialization(
        &mut self, 
        val: impl ValueView
    ) -> PartialVMResult<()>;
    fn charge_copy_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>;
    fn charge_move_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>;
    fn charge_store_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>;
    fn charge_pack(
        &mut self, 
        is_generic: bool, 
        args: impl ExactSizeIterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_unpack(
        &mut self, 
        is_generic: bool, 
        args: impl ExactSizeIterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_read_ref(&mut self, val: impl ValueView) -> PartialVMResult<()>;
    fn charge_write_ref(
        &mut self, 
        new_val: impl ValueView, 
        old_val: impl ValueView
    ) -> PartialVMResult<()>;
    fn charge_eq(
        &mut self, 
        lhs: impl ValueView, 
        rhs: impl ValueView
    ) -> PartialVMResult<()>;
    fn charge_neq(
        &mut self, 
        lhs: impl ValueView, 
        rhs: impl ValueView
    ) -> PartialVMResult<()>;
    fn charge_borrow_global(
        &mut self, 
        is_mut: bool, 
        is_generic: bool, 
        ty: impl TypeView, 
        is_success: bool
    ) -> PartialVMResult<()>;
    fn charge_exists(
        &mut self, 
        is_generic: bool, 
        ty: impl TypeView, 
        exists: bool
    ) -> PartialVMResult<()>;
    fn charge_move_from(
        &mut self, 
        is_generic: bool, 
        ty: impl TypeView, 
        val: Option<impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_move_to(
        &mut self, 
        is_generic: bool, 
        ty: impl TypeView, 
        val: impl ValueView, 
        is_success: bool
    ) -> PartialVMResult<()>;
    fn charge_vec_pack<'a>(
        &mut self, 
        ty: impl TypeView + 'a, 
        args: impl ExactSizeIterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_vec_len(&mut self, ty: impl TypeView) -> PartialVMResult<()>;
    fn charge_vec_borrow(
        &mut self, 
        is_mut: bool, 
        ty: impl TypeView, 
        is_success: bool
    ) -> PartialVMResult<()>;
    fn charge_vec_push_back(
        &mut self, 
        ty: impl TypeView, 
        val: impl ValueView
    ) -> PartialVMResult<()>;
    fn charge_vec_pop_back(
        &mut self, 
        ty: impl TypeView, 
        val: Option<impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_vec_unpack(
        &mut self, 
        ty: impl TypeView, 
        expect_num_elements: NumArgs, 
        elems: impl ExactSizeIterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_vec_swap(&mut self, ty: impl TypeView) -> PartialVMResult<()>;
    fn charge_load_resource(
        &mut self, 
        loaded: Option<(NumBytes, impl ValueView)>
    ) -> PartialVMResult<()>;
    fn charge_native_function(
        &mut self, 
        amount: InternalGas, 
        ret_vals: Option<impl ExactSizeIterator<Item = impl ValueView>>
    ) -> PartialVMResult<()>;
    fn charge_native_function_before_execution(
        &mut self, 
        ty_args: impl ExactSizeIterator<Item = impl TypeView>, 
        args: impl ExactSizeIterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;
    fn charge_drop_frame(
        &mut self, 
        locals: impl Iterator<Item = impl ValueView>
    ) -> PartialVMResult<()>;

}

Trait that defines a generic gas meter interface, allowing clients of the Move VM to implement their own metering scheme.

Required Methods

fn charge_simple_instr(
&mut self,
instr: SimpleInstruction
) -> PartialVMResult<()>

Charge an instruction and fail if not enough gas units are left.

fn charge_pop(&mut self, popped_val: impl ValueView) -> PartialVMResult<()>

fn charge_call(
&mut self,
module_id: &ModuleId,
func_name: &str,
args: impl ExactSizeIterator<Item = impl ValueView>,
num_locals: NumArgs
) -> PartialVMResult<()>

fn charge_call_generic(
&mut self,
module_id: &ModuleId,
func_name: &str,
ty_args: impl ExactSizeIterator<Item = impl TypeView>,
args: impl ExactSizeIterator<Item = impl ValueView>,
num_locals: NumArgs
) -> PartialVMResult<()>

fn charge_ld_const(&mut self, size: NumBytes) -> PartialVMResult<()>

fn charge_ld_const_after_deserialization(
&mut self,
val: impl ValueView
) -> PartialVMResult<()>

fn charge_copy_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>

fn charge_move_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>

fn charge_store_loc(&mut self, val: impl ValueView) -> PartialVMResult<()>

fn charge_pack(
&mut self,
is_generic: bool,
args: impl ExactSizeIterator<Item = impl ValueView>
) -> PartialVMResult<()>

fn charge_unpack(
&mut self,
is_generic: bool,
args: impl ExactSizeIterator<Item = impl ValueView>
) -> PartialVMResult<()>

fn charge_read_ref(&mut self, val: impl ValueView) -> PartialVMResult<()>

fn charge_write_ref(
&mut self,
new_val: impl ValueView,
old_val: impl ValueView
) -> PartialVMResult<()>

fn charge_eq(
&mut self,
lhs: impl ValueView,
rhs: impl ValueView
) -> PartialVMResult<()>

fn charge_neq(
&mut self,
lhs: impl ValueView,
rhs: impl ValueView
) -> PartialVMResult<()>

fn charge_borrow_global(
&mut self,
is_mut: bool,
is_generic: bool,
ty: impl TypeView,
is_success: bool
) -> PartialVMResult<()>

fn charge_exists(
&mut self,
is_generic: bool,
ty: impl TypeView,
exists: bool
) -> PartialVMResult<()>

fn charge_move_from(
&mut self,
is_generic: bool,
ty: impl TypeView,
val: Option<impl ValueView>
) -> PartialVMResult<()>

fn charge_move_to(
&mut self,
is_generic: bool,
ty: impl TypeView,
val: impl ValueView,
is_success: bool
) -> PartialVMResult<()>

fn charge_vec_pack<'a>(
&mut self,
ty: impl TypeView + 'a,
args: impl ExactSizeIterator<Item = impl ValueView>
) -> PartialVMResult<()>

fn charge_vec_len(&mut self, ty: impl TypeView) -> PartialVMResult<()>

fn charge_vec_borrow(
&mut self,
is_mut: bool,
ty: impl TypeView,
is_success: bool
) -> PartialVMResult<()>

fn charge_vec_push_back(
&mut self,
ty: impl TypeView,
val: impl ValueView
) -> PartialVMResult<()>

fn charge_vec_pop_back(
&mut self,
ty: impl TypeView,
val: Option<impl ValueView>
) -> PartialVMResult<()>

fn charge_vec_unpack(
&mut self,
ty: impl TypeView,
expect_num_elements: NumArgs,
elems: impl ExactSizeIterator<Item = impl ValueView>
) -> PartialVMResult<()>

fn charge_vec_swap(&mut self, ty: impl TypeView) -> PartialVMResult<()>

fn charge_load_resource(
&mut self,
loaded: Option<(NumBytes, impl ValueView)>
) -> PartialVMResult<()>

Charges for loading a resource from storage. This is only called when the resource is not cached.

Some(n) means n bytes are loaded.

None means a load operation is performed but the resource does not exist.

WARNING: This can be dangerous if you execute multiple user transactions in the same session – identical transactions can have different gas costs. Use at your own risk.

fn charge_native_function(
&mut self,
amount: InternalGas,
ret_vals: Option<impl ExactSizeIterator<Item = impl ValueView>>
) -> PartialVMResult<()>

Charge for executing a native function. The cost is calculated returned by the native function implementation. Should fail if not enough gas units are left.

In the future, we may want to remove this and directly pass a reference to the GasMeter instance to the native functions to allow gas to be deducted during computation.

fn charge_native_function_before_execution(
&mut self,
ty_args: impl ExactSizeIterator<Item = impl TypeView>,
args: impl ExactSizeIterator<Item = impl ValueView>
) -> PartialVMResult<()>

fn charge_drop_frame(
&mut self,
locals: impl Iterator<Item = impl ValueView>
) -> PartialVMResult<()>