Module move_vm_types::gas_schedule

This module lays out the basic abstract costing schedule for bytecode instructions.

It is important to note that the cost schedule defined in this file does not track hashing operations or other native operations; the cost of each native operation will be returned by the native function itself.

Structs

GasStatus
The Move VM implementation of state for gas metering.

Enums

NativeCostIndex

Statics

INITIAL_GAS_SCHEDULE

Functions

calculate_intrinsic_gas
Calculate the intrinsic gas for the transaction based upon its size in bytes/words.

new_from_instructions

zero_cost_schedule