Struct move_vm_types::gas_schedule::CostStrategy

pub struct CostStrategy<'a> { /* fields omitted */ }

The Move VM implementation for gas charging.

Initialize with a CostTable and the gas provided to the transaction. Provide all the proper guarantees about gas charging in the Move VM.

Every client must use an instance of this type to interact with the Move VM.

Methods

impl<'a> CostStrategy<'a>

pub fn transaction(
cost_table: &'a CostTable,
gas_left: GasUnits<GasCarrier>
) -> Self

A transaction CostStrategy. Charge for every operation and fails once there is no more gas to pay for operations.

This is the instantiation the must be used when execution a user script.

pub fn system(cost_table: &'a CostTable, gas_left: GasUnits<GasCarrier>) -> Self

A system CostStrategy does not charge for operations.

It should be used by clients in very specific cases and when executing system code that does not have to charge the user.

pub fn cost_table(&self) -> &CostTable

Return the CostTable behind this CostStrategy.

pub fn remaining_gas(&self) -> GasUnits<GasCarrier>

Return the gas left.

pub fn deduct_gas(&mut self, amount: GasUnits<GasCarrier>) -> VMResult<()>

Charge a given amount of gas and fail if not enough gas units are left.

pub fn charge_instr_with_size(
&mut self,
opcode: Opcodes,
size: AbstractMemorySize<GasCarrier>
) -> VMResult<()>

Charge an instruction over data with a given size and fail if not enough gas units are left.

pub fn charge_instr(&mut self, opcode: Opcodes) -> VMResult<()>

Charge an instruction and fail if not enough gas units are left.

pub fn charge_intrinsic_gas(
&mut self,
size: AbstractMemorySize<GasCarrier>
) -> VMResult<()>

Charge gas related to the overall size of a transaction and fail if not enough gas units are left.