Module move_vm_types::loaded_data::types

Loaded representation for runtime types.

Modules

prop

Structs

FatStructType
VM representation of a struct type in Move.

Enums

FatType
VM representation of a Move type that gives access to both the fully qualified name and data layout of the type.