Enum move_vm_types::loaded_data::types::FatType

pub enum FatType {
    Bool,
    U8,
    U64,
    U128,
    Address,
    Signer,
    Vector(Box<FatType>),
    Struct(Box<FatStructType>),
    Reference(Box<FatType>),
    MutableReference(Box<FatType>),
    TyParam(usize),
}

VM representation of a Move type that gives access to both the fully qualified name and data layout of the type.

TODO: this data structure itself is intended to be used in runtime only and should NOT be serialized in any form. Currently we still derive Serialize and Deserialize, but this is a hack for fuzzing and should be guarded behind the "fuzzing" feature flag. We should look into ways to get rid of this.

Variants

Bool

U8

U64

U128

Address

Signer

Vector(Box<FatType>)

Struct(Box<FatStructType>)

Reference(Box<FatType>)

MutableReference(Box<FatType>)

TyParam(usize)

Methods

impl FatType

pub fn single_value_strategy() -> impl Strategy<Value = Self>

Generate a random primitive Type, no Struct or Vector.

pub fn nested_strategy(
depth: u32,
desired_size: u32,
expected_branch_size: u32
) -> impl Strategy<Value = Self>

Generate a primitive Value, a Struct or a Vector.

impl FatType

pub fn subst(&self, ty_args: &[FatType]) -> VMResult<FatType>

pub fn type_tag(&self) -> VMResult<TypeTag>

pub fn is_resource(&self) -> VMResult<bool>

pub fn debug_print<B: Write>(&self, buf: &mut B) -> VMResult<()>