Macro move_vm_types::pop_arg

macro_rules! pop_arg {
    ($arguments:ident, $t:ty) => { ... };
}

Return the argument at the top of the stack.

Arguments are passed to a native as a stack with first arg at the bottom of the stack. Calling this API can help in making the code more readable. It's good practice to pop all arguments in locals of the native function on function entry.