Trait move_vm_types::natives::function::NativeContext

pub trait NativeContext {
    fn print_stack_trace<B: Write>(&self, buf: &mut B) -> PartialVMResult<()>;

    fn cost_table(&self) -> &CostTable;

    fn save_event(
        &mut self, 
        guid: Vec<u8>, 
        count: u64, 
        ty: Type, 
        val: Value
    ) -> PartialVMResult<bool>;

    fn type_to_type_layout(
        &self, 
        ty: &Type
    ) -> PartialVMResult<Option<MoveTypeLayout>>;
}

NativeContext - Native function context.

This is the API, the “privileges”, a native function is given. Normally a native function will only need the CostTable. The set of native functions and their linkage is entirely inside the MoveVM runtime.

Required methods

fn print_stack_trace<B: Write>(&self, buf: &mut B) -> PartialVMResult<()>

Prints stack trace.

fn cost_table(&self) -> &CostTable

Gets cost table ref.

fn save_event(
&mut self,
guid: Vec<u8>,
count: u64,
ty: Type,
val: Value
) -> PartialVMResult<bool>

Saves contract event. Returns true if successful

fn type_to_type_layout(
&self,
ty: &Type
) -> PartialVMResult<Option<MoveTypeLayout>>

Get the a data layout via the type.