Struct move_vm_types::values::values_impl::GlobalValue

pub struct GlobalValue { /* fields omitted */ }

A special value that lives in global storage.

Callers are allowed to take global references from a GlobalValue. A global value also contains an internal flag, indicating whether the value has potentially been modified or not.

For any given value in storage, only one GlobalValue may exist to represent it at any time. This means that:

GlobalValue does not and cannot implement Clone!

a borrowed reference through borrow_global is represented through a &GlobalValue.

borrow_global_mut is also represented through a &GlobalValue -- the bytecode verifier enforces mutability restrictions.

move_from is represented through an owned GlobalValue.

Methods

impl GlobalValue

pub fn size(&self) -> AbstractMemorySize<GasCarrier>

impl GlobalValue

pub fn new(v: Value) -> VMResult<Self>

pub fn borrow_global(&self) -> VMResult<Value>

pub fn mark_dirty(&self) -> VMResult<()>

pub fn is_clean(&self) -> VMResult<bool>

pub fn is_dirty(&self) -> VMResult<bool>

pub fn into_owned_struct(self) -> VMResult<Struct>