Trait move_vm_types::views::ValueView

pub trait ValueView {
    fn visit(&self, visitor: &mut impl ValueVisitor);

    fn legacy_abstract_memory_size(&self) -> AbstractMemorySize { ... }
}

Trait that provides an abstract view into a Move Value.

This is used to expose certain info to clients (e.g. the gas meter), usually in a lazily evaluated fashion.

Required Methods

fn visit(&self, visitor: &mut impl ValueVisitor)

Provided Methods

fn legacy_abstract_memory_size(&self) -> AbstractMemorySize

Returns the abstract memory size of the value.

The concept of abstract memory size is not well-defined and is only kept for backward compatibility. New applications should avoid using this.

TODO(Gas): Encourage clients to replicate this in their own repo and get this removed once they are done.