Trait move_vm_types::data_store::DataStore

pub trait DataStore {
    fn publish_resource(
        &mut self, 
        ap: &AccessPath, 
        g: (FatStructType, GlobalValue)
    ) -> VMResult<()>;

    fn borrow_resource(
        &mut self, 
        ap: &AccessPath, 
        ty: &FatStructType
    ) -> VMResult<Option<&GlobalValue>>;

    fn move_resource_from(
        &mut self, 
        ap: &AccessPath, 
        ty: &FatStructType
    ) -> VMResult<Option<GlobalValue>>;

    fn load_module(&self, module: &ModuleId) -> VMResult<Vec<u8>>;

    fn publish_module(
        &mut self, 
        module_id: ModuleId, 
        module: Vec<u8>
    ) -> VMResult<()>;

    fn exists_module(&self, key: &ModuleId) -> bool;

    fn emit_event(&mut self, event: ContractEvent);
}

Provide an implementation for bytecodes related to data with a given data store.

The DataStore is a generic concept that includes both data and events. A default implementation of the DataStore is TransactionDataCache which provides an in memory cache for a given transaction and the atomic transactional changes proper of a script execution (transaction).

Required methods

fn publish_resource(
&mut self,
ap: &AccessPath,
g: (FatStructType, GlobalValue)
) -> VMResult<()>

Publish a resource.

fn borrow_resource(
&mut self,
ap: &AccessPath,
ty: &FatStructType
) -> VMResult<Option<&GlobalValue>>

Get a reference to a resource.

fn move_resource_from(
&mut self,
ap: &AccessPath,
ty: &FatStructType
) -> VMResult<Option<GlobalValue>>

Transfer ownership of a resource to the VM.

fn load_module(&self, module: &ModuleId) -> VMResult<Vec<u8>>

Get the serialized format of a CompiledModule given a ModuleId.

fn publish_module(
&mut self,
module_id: ModuleId,
module: Vec<u8>
) -> VMResult<()>

Publish a module.

fn exists_module(&self, key: &ModuleId) -> bool

Check if this module exists.

fn emit_event(&mut self, event: ContractEvent)

Emit an event to the EventStore

Loading content...