Module move_vm_types::natives::function

Native Function Support

All move native functions have the following signature:

pub fn native_function( context: &mut impl NativeContext, ty_args: Vec<Type>, mut arguments: VecDeque<Value>, ) -> VMResult<NativeResult>;

arguments are passed with first argument at position 0 and so forth. Popping values from arguments gives the aguments in reverse order (last first). This module contains the declarations and utilities to implement a native function.

Structs

NativeResult
Result of a native function execution requires charges for execution cost.

Traits

NativeContext
NativeContext - Native function context.

Functions

native_gas
Return the native gas entry in CostTable for the given key. The key is the specific native function index known to CostTable.