Trait move_vm_types::natives::function::NativeContext

pub trait NativeContext {
    fn print_stack_trace<B: Write>(&self, buf: &mut B) -> VMResult<()>;

    fn cost_table(&self) -> &CostTable;

    fn save_under_address(
        &mut self, 
        ty_args: &[Type], 
        module_id: &ModuleId, 
        struct_name: &IdentStr, 
        resource_to_save: Struct, 
        account_address: AccountAddress
    ) -> VMResult<()>;

    fn save_event(&mut self, event: ContractEvent) -> VMResult<()>;

    fn convert_to_fat_types(&self, types: Vec<Type>) -> VMResult<Vec<FatType>>;
}

NativeContext - Native function context.

This is the API, the "privileges", a native function is given. Normally a native function will only need the CostTable. The set of native functions and their linkage is entirely inside the MoveVM runtime.

Required methods

fn print_stack_trace<B: Write>(&self, buf: &mut B) -> VMResult<()>

Prints stack trace.

fn cost_table(&self) -> &CostTable

Gets cost table ref.

fn save_under_address(
&mut self,
ty_args: &[Type],
module_id: &ModuleId,
struct_name: &IdentStr,
resource_to_save: Struct,
account_address: AccountAddress
) -> VMResult<()>

fn save_event(&mut self, event: ContractEvent) -> VMResult<()>

Saves contract event.

fn convert_to_fat_types(&self, types: Vec<Type>) -> VMResult<Vec<FatType>>

Converts types to fet types.

Loading content...