Module move_vm_types::values::values_impl

Modules

debug

prop

Structs

GlobalValue
A wrapper around GlobalValueImpl, representing a “slot” in global storage that can hold a resource.

Locals
The locals for a function frame. It allows values to be read, written or taken reference from.

Reference
A generic Move reference that offers two functionalities: read_ref & write_ref.

SignerRef

Struct
A Move struct.

StructRef
A reference to a Move struct that allows you to take a reference to one of its fields.

Value
A Move value – a wrapper around ValueImpl which can be created only through valid means.

Vector

VectorRef

Enums

GlobalValueEffect
Simple enum for the change state of a GlobalValue, used by into_effect.

IntegerValue
An integer value in Move.

Constants

DESTROY_NON_EMPTY_VEC

INDEX_OUT_OF_BOUNDS

POP_EMPTY_VEC

Traits

VMValueCast