Module move_vm_types::values::values_impl

Modules

debug

prop

vector

Structs

GlobalValue
A special value that lives in global storage.

Locals
The locals for a function frame. It allows values to be read, written or taken reference from.

Reference
A generic Move reference that offers two functinalities: read_ref & write_ref.

Struct
A Move struct.

StructRef
A reference to a Move struct that allows you to take a reference to one of its fields.

Value
A Move value -- a wrapper around ValueImpl which can be created only through valid means.

Enums

IntegerValue
An integer value in Move.

Traits

VMValueCast